{{ cookiecutter.project_name }}/requirements/toolbox.txt
# A set of common tools to use in local development only
cookiecutter>=2.0,<3.0
pipdeptree
django-extensions
django-debug-toolbar
